Code: Split the mode control functions into smaller 'layered' functions (HOP-6266)
Problem: During breaking down of a mode, some (higher level) plugins react to the lower level config being changed.

Cause: The configuration for the full IPv4Mode gets broken down before the IPv6Mode is started. If they have a joint lower layer this means the higher level IPv6Mode config is still inplace when the lower level is removed.

Solution: Instead of configuring/breaking down the whole IPv4Mode before the IPv6Mode, the mode control functions are split in layers. This allows to call both the IPv4 and IPv6 layer functions before continuing to the next layer.
As an added benifit common layers can be reused between modes and do not need to be addapted for each mode seperatly.

Issue: HOP-6266 - Rework mode toggling to a layer approach
GitOrigin-RevId: 0b255aec2e2581b85bc918f64da84a77518ecdd8