unit-tests: Update unit-tests to new development (HOP-6266)
Problem: Unit-tests do not pass after the rework to a layered toggle.

Cause: The order is now always IPv4 first and IPv6 second. So the order of the lowerlayers switched as before it sometimes could be IPv6 first.

Solution: Change the order of the lowerlayers to match.

Issue: HOP-6266 - Rework mode toggling to a layer approach
GitOrigin-RevId: bed1986e527d46ca3a7c064ce21ed413f8f58023