Query cleanup: Clean up all the queries on intf delete (HOP-7708/PPW-106)
Problem: When deleting an intf instance from a wan-mode the wan-manager crashes

Cause: When an interface is deleted, it uses the objects priv data as if it is a single query.
This is not the case, it is a struct containing multiple queries.
Trying to close it will result in a crash.

Solution: Cast the priv to the proper type and call the intf_isup_queries_clean to make sure the data is properly cleaned.

Issue: HOP-7708 - Static wan deletion from datamodel crashes in libnetmodel
Issue: PPW-106 - Static wan deletion from datamodel crashes in libnetmodel

Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>
GitOrigin-RevId: f42de13905ee2249e8e48190b2027a4d19429f66