PROVISIONAL: Remove dependency of mod-wanmgr-sfp
In wan-manager version v1.6.19 support was added for additional modules.
These modules of which mod-wanmgr-sfp is one of cannot be a dependency of wan-manager.

The idea of the modules is that they can be optionally enabled to add more functionality.
This means that when the module is not enabled in the datamodel the plugin still has to work correctly,
which is not the case, as a consequence a WANMode cannot be set:
wan-manager: wanmgr_ - [x]Failed to execute read-sfpcategory function, '-1' - (read_sfp_category@sfp/sfp.c:329)
wan-manager: wan-man - [x]Unable to read sfp_category, '-1' - (wan_manager_found_ll@dm_wan_mode.c:272)

!!! This is a temporary fix and should be properly fixed

References: PCF-1750
Signed-off-by: Sam Maes <sam.maes_ext@softathome.com>