PROVISIONAL: Remove dependency of mod-wanmgr-system
In wan-manager version v1.6.19 support was added for additional modules.
These modules of which mod-wanmgr-system is one of cannot be a dependency of wan-manager.

The idea of the modules is that they can be optionally enabled to add more functionality.
This means that when the module is not enabled in the datamodel the plugin still has to work correctly,
which is not the case, as a consequence a WANMode cannot be set:
wan-manager: wan-man - [x]Request system update failed, rc = -1 - (wan_mode_set@dm_wan_mode.c:593)
wan-manager: as-ctrl - [x]Failed to set wan mode 'demo_wanmode' - (autosensing_set_wan_mode@autosensing/autosensing.c:120)

!!! This is a temporary fix and should be properly fixed

References: PCF-1750
Signed-off-by: Sam Maes <sam.maes_ext@softathome.com>