get_ip_path: Only use Device prefixed path when needed
It is faster to use the direct path to the IP-manager when possible to not go through the DM-proxy.
To make this possible, get_ip_path is changed to allow returning both prefixed and none prefixed paths.

All calls setting in or getting from the IP-manager should use none device prefixed paths.
Everywhere the path is used to fill in a path in the datamodel, device prefixed paths shoul be used.

References: PPW-521, PPW-522
Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>