get_{x}_path: Return none device prefixed paths
When getting or setting data from/on a component, it is faster to use the none Device prefixed path to not go through the DM-proxy.
To do this, all the get_{x}_path functions need to be changed to remove the Device part from the path.

The DHCPv4 and DHCPv6 paths are special as they can not be converted by just removing the Device parth so they have a replace function instead.

References: PPW-521, PPW-522
Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>