Remove use of DEVICE_PATH define where no longer needed
Some functions use a fixed path to get or set data in a component.
The paths are often created by using the DEVICE_PATH definition in from of the normal device path.

As it is faster to use the direct path to get and set information from a component,
the DEVICE_PATH can be removed from a lot of path creations to simplify the operation.

References: PPW-521, PPW-522
Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>