PROVISORY: odl extensions: Remove mod-wanmgr-system in the default
mod-wanmgr-system is a module to handle vendor specific requirements based on the WAN change to update the L1 driver settings and reboot the board if needed.

The current fast track boards (Freedom and OSPv2) don't require a reboot on a wan change but the current module does perform one. This means the board is unnecessarily rebooted with some wan changes.

To prevent this behavior, we should not include the module by default and only include the module for boards where a reboot is needed.

References: PPW-575
Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>