unit-test: test_start_sensing.json: Remove mod-wanmgr-system from expected value
After removing the mod-wanmgr-system as a default module from the wan-manager, unit tests fail.
This is because the tests still expect the ExtensionModules still to include the mod-wanmgr-system in its value.

The expected output should be updated to not include the this module to be in line again with the current default.

References: PPW-575

Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>