wan-manager: routing: add ipv6 route information support
Routing.RouteInformation.InterfaceSetting

In the routing-manager, the implementation of the data model has been
updated to be in line with the TR1-181 specification. Several parameters
and objects are declared read-only now. One of these is the template-
object Routing.RouteInformation.InterfaceSetting. The wan-manager can
thus no longer create instances of this object in case of a wan-mode
with multiple wan interfaces (e.g. modes with a failover interface).
It is now up to the routing-manager to create all the necessary
InterfaceSetting instances. The wan-manager shouldn't do anything
with regards to this, so all references to Routing.RouteInformation
are removed.

Reference: PCF-1850

Signed-off-by: Benamor Imene <imene.benamor_ext@softathome.com>